[Update] Finish server
