+++ ChatAppServer.WebAPI/bin/Debug/net8.0/ChatAppServer.WebAPI.staticwebassets.runtime.json
@@ -1,1 +1,1 @@
-{"ContentRoots":["D:\\Harmony-Chat\\ChatAppServer.WebAPI\\wwwroot\\"],"Root":{"Children":{"avatars":{"Children":{"default.jpg":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"avatars/default.jpg"},"Patterns":null}},"Asset":null,"Patterns":null}},"Asset":null,"Patterns":[{"ContentRootIndex":0,"Pattern":"**","Depth":0}]}}
+{"ContentRoots":["B:\\AD\\work\\.net\\Harmony-Chat\\ChatAppServer.WebAPI\\wwwroot\\"],"Root":{"Children":{"avatars":{"Children":{"default.jpg":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"avatars/default.jpg"},"Patterns":null}},"Asset":null,"Patterns":null}},"Asset":null,"Patterns":[{"ContentRootIndex":0,"Pattern":"**","Depth":0}]}}
