+++ ChatAppServer.WebAPI/bin/Debug/net8.0/appsettings.json
@@ -4,7 +4,7 @@
     "Issuer": "ChatAppServer"
   },
   "ConnectionStrings": {
-    "MySqlConnection": "Server=localhost;Database=ChatAppDb;User=root;Password=Pha9124719@;Connect Timeout=30;"
+    "MySqlConnection": "Server=localhost;Database=ChatAppDb;User=root;Password=;Connect Timeout=30;"
   },
   "Smtp": {
     "Host": "smtp.gmail.com",
